Changed target and claim id to smallint datatype.

+++ reporting/sql/V201702061486427077__initial_ddl.sql
@@ -106,7 +106,7 @@ CREATE TABLE IF NOT EXISTS asmt_score (
 );
 
 CREATE TABLE IF NOT EXISTS claim (
-  id int NOT NULL PRIMARY KEY,
+  id smallint NOT NULL PRIMARY KEY,
   subject_id tinyint NOT NULL,
   code varchar(10) NOT NULL,
   name varchar(250) NOT NULL,
@@ -125,8 +125,8 @@ CREATE TABLE IF NOT EXISTS subject_claim_score (
 );
 
 CREATE TABLE IF NOT EXISTS target (
-  id int NOT NULL PRIMARY KEY,
-  claim_id int NOT NULL,
+  id smallint NOT NULL PRIMARY KEY,
+  claim_id smallint NOT NULL,
   code varchar(10) NOT NULL,
   description varchar(500) NOT NULL,
   CONSTRAINT fk__target__claim FOREIGN KEY (claim_id) REFERENCES claim(id)
@@ -148,8 +148,8 @@ CREATE TABLE IF NOT EXISTS math_practice (
 
 CREATE TABLE IF NOT EXISTS item (
   id int NOT NULL PRIMARY KEY,
-  claim_id int,
-  target_id int,
+  claim_id smallint,
+  target_id smallint,
   natural_id varchar(40) NOT NULL,
   asmt_id int NOT NULL,
   math_practice tinyint,
